fix: folders with names of unlinked entries are linked (#1027)
* fix: render folder entries as unlinked

* fix: hide file size for linked folders

* fix: linked folders can be revealed in explorer

* fix: linked folders can be opened

* fix: linked folders can be deleted

* fix: skip rendering thumbnails in `ThumbRenderer._render`

* fix: skip getting image metadata for folders

* fix: conflicts

* style: ruff

+++ src/tagstudio/qt/helpers/file_deleter.py
@@ -21,6 +21,8 @@ def delete_file(path: str | Path) -> bool:
     """
     _path = Path(path)
     try:
+        if _path.is_dir():
+            return False
         logger.info(f"[delete_file] Sending to Trash: {_path}")
         send2trash(_path)
         return True
